Round up percentage

+++ plugins/ai.py
@@ -112,7 +112,7 @@ class Ai(object):
 
             # Percentage of global lines the current channel accounts for.
             if channel_line_count >= 0 and line_count >= 0:
-                channel_percentage = 100 * float(channel_line_count) / float(line_count)
+                channel_percentage = int(100 * float(channel_line_count) / float(line_count))
 
             return '{0}: Chatbot is currently {1} for {4}. Channel/global line count: {3}/{2} ({5}%).'.format(
                 mask.nick, 'enabled' if self.is_active(target) else 'disabled',
